src/otel-python-app/requirements.txt
Flask>=3.0.0,<4.0.0
opentelemetry-distro>=0.44b0
opentelemetry-exporter-otlp-proto-http>=1.21.0
opentelemetry-instrumentation-flask>=0.44b0
